dialog/requirements.txt
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comtypes==1.4.13
filelock==3.20.0
fsspec==2025.10.0
gTTS==2.5.4
idna==3.11
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
openai-whisper==20250625
paho-mqtt==2.1.0
PyAudio==0.2.14
pydub==0.25.1
# Windows-specific dependencies (Conditional Installation)
pywin32>=223 ; platform_system=="Windows"
pypiwin32 ; platform_system=="Windows"
pyttsx3==2.91
regex==2025.11.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
SpeechRecognition==3.14.3
sympy==1.14.0
termcolor==3.2.0
tiktoken==0.12.0
torch==2.9.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
whisper==1.1.10
